+++ tests/testthat/test-estimate_cate.R
@@ -21,6 +21,7 @@ test_that("CATE (DRLearner) Estimation Runs Correctly", {
   node_size <- 20
   max_nodes <- 5
   max_depth <- 15
+  replace <- FALSE
   max_decay <- 0.025
   type_decay <- 2
   t <- 0.025
@@ -67,7 +68,7 @@ test_that("CATE (DRLearner) Estimation Runs Correctly", {
 
   # Step 3: Generate rules list
   initial_rules_dis <- generate_rules(X_dis, ite_std_dis, ntrees_rf, ntrees_gbm,
-                                      node_size, max_nodes, max_depth,
+                                      node_size, max_nodes, max_depth, replace,
                                       random_state = 981)
 
   rules_list_dis <- prune_rules(initial_rules_dis, X_dis, ite_std_dis, max_decay, type_decay)
@@ -275,6 +276,7 @@ test_that("CATE (cf-means) Estimation Runs Correctly", {
   node_size <- 20
   max_nodes <- 5
   max_depth <- 15
+  replace <- FALSE
   max_decay <- 0.025
   type_decay <- 2
   t <- 0.025
@@ -318,7 +320,7 @@ test_that("CATE (cf-means) Estimation Runs Correctly", {
   # Generate rules list
   initial_rules_dis <- CRE:::generate_rules(X_dis, ite_std_dis, ntrees_rf,
                                             ntrees_gbm, node_size, max_nodes,
-                                            max_depth, random_state = 214)
+                                            max_depth, replace, random_state = 214)
 
   rules_list_dis <- CRE:::prune_rules(initial_rules_dis, X_dis, ite_std_dis, max_decay, type_decay)