+++ tests/testthat/test-interpret_select_rules.R
@@ -15,6 +15,7 @@ test_that("Rules Interpreted Correctly", {
   node_size <- 20
   max_nodes <- 5
   max_depth = 15
+  replace <- FALSE
   max_decay = 0.025
   type_decay = 2
   t <- 0.025
@@ -41,7 +42,7 @@ test_that("Rules Interpreted Correctly", {
 
   # Step 3: Generate rules list
   initial_rules <- generate_rules(X, ite_std, ntrees_rf, ntrees_gbm, node_size,
-                                  max_nodes, max_depth, random_state = 812)
+                                  max_nodes, max_depth, replace, random_state = 812)
 
   rules_list <- prune_rules(initial_rules, X, ite_std, max_decay, type_decay)